+++ src/const.ts
@@ -1,0 +1,9 @@
+
+export const DEFAULT_MAX_RESULTS = 10
+
+
+export const MENUS = {
+  CREATE: 'create',
+  LIST: 'list',
+  EXIT: 'exit',
+}
